requirements.txt
python-dotenv==1.1.0

langchain
langchain-openai
langchain-community
langgraph

# tools
duckduckgo-search
langchain-tavily	
langchain-experimental

playwright
lxml
requests
faiss-cpu
pandas[excel]
youtube-transcript-api
yt-dlp
openai

stockfish
pint==0.24.4
markdownify
pdfplumber
python-pptx==1.0.2
mammoth